refactor: Organisation
