+++ exercise/C#/day12/Gifts/Gifts/Domain/Behaviors/NaughtyBehavior.cs
@@ -1,6 +1,6 @@
 using LanguageExt;
 
-namespace Gifts;
+namespace Gifts.Domain.Behaviors;
 
 internal class NaughtyBehavior : Behavior
 {
