+++ exercise/C#/day12/Gifts/Gifts/Domain/Behaviors/NiceBehavior.cs
@@ -1,6 +1,6 @@
 using LanguageExt;
 
-namespace Gifts;
+namespace Gifts.Domain.Behaviors;
 
 internal class NiceBehavior : Behavior
 {
