+++ exercise/C#/day12/Gifts/Gifts/Domain/Behaviors/VeryNiceBehavior.cs
@@ -1,6 +1,6 @@
 using LanguageExt;
 
-namespace Gifts;
+namespace Gifts.Domain.Behaviors;
 
 internal class VeryNiceBehavior : Behavior
 {
