+++ exercise/C#/day12/Gifts/Gifts/Domain/WishList.cs
@@ -1,6 +1,6 @@
 using LanguageExt;
 
-namespace Gifts;
+namespace Gifts.Domain;
 
 internal class WishList
 {
