+++ exercise/C#/day12/Gifts/Gifts/Ports/IChildrenRepository.cs
@@ -1,6 +1,7 @@
+using Gifts.Domain;
 using LanguageExt;
 
-namespace Gifts;
+namespace Gifts.Ports;
 
 public interface IChildrenRepository
 {
